requirements.txt
filelock==3.16.1
fsspec==2024.12.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
pillow==10.4.0
scikit-learn==1.3.2
scipy==1.10.1
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.4.1
torchvision==0.19.1
typing_extensions==4.12.2
